requirements.txt
absl-py==1.4.0
alembic==1.8.1
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.1.0
astunparse==1.6.3
attrs==22.1.0
autocommand==2.2.2
autopage==0.5.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==4.1.0
bleach==5.0.1
blinker==1.6.2
cachetools==5.3.0
celery==5.3.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cliff==4.0.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.5
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.5
cryptography==41.0.4
cycler==0.11.0
deap==1.3.3
debugpy==1.6.6
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.14
docx==0.2.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.3.3
fonttools==4.37.4
fqdn==1.5.1
freezegun==1.2.2
fsspec==2023.6.0
gast==0.4.0
gensim==4.3.1
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==4.13.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.17.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==1.5.1
jupyter-nbextensions-configurator==0.6.1
jupyter-server==1.23.5
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server_terminals==0.4.2
jupyterlab==3.5.3
jupyterlab-lsp==3.10.2
jupyterlab-pygments==0.2.2
jupyterlab-tabnine==0.0.24
jupyterlab-widgets==3.0.4
jupyterlab_server==2.19.0
kaggle==1.5.12
kaleido==0.2.1
keras==2.12.0
kiwisolver==1.4.4
kombu==5.3.2
libclang==16.0.0
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mecab==0.996.3
mecab-python3==1.0.6
mistune==2.0.4
ml-dtypes==0.0.4
MouseInfo==0.1.3
mpmath==1.2.1
MultiApp==1.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
oathtool==2.3.1
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.0.3
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.7.1
path==16.6.0
pbr==5.11.0
pickleshare==0.7.5
Pillow==9.2.0
plac==1.3.5
platformdirs==2.6.0
plotly==5.15.0
pluggy==1.0.0
prettytable==3.4.1
prometheus-client==0.15.0
prompt-toolkit==3.0.39
protobuf==4.22.1
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pydot==1.4.2
pygame==2.3.0
PyGetWindow==0.0.9
PyGithub==2.1.1
Pygments==2.13.0
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.19.1
python-json-logger==2.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-pptx==0.6.21
python-slugify==6.1.2
pytube==15.0.0
pytweening==1.0.4
pytz==2022.4
pywin32==304
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.8.8
reportlab==4.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.2
scikit-learn==1.1.2
scipy==1.8.1
seaborn==0.12.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==6.3.0
sniffio==1.3.0
soundfile==0.11.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
stack-data==0.6.3
stevedore==4.1.0
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torch-tb-profiler==0.4.1
torchaudio==2.0.2
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.64.1
traitlets==5.12.0
transformers==4.31.0
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.7.0
unidic==1.1.0
uri-template==1.2.0
urllib3==1.26.0
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.8
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.3.6
wfdb==4.1.0
widgetsnbextension==4.0.4
wrapt==1.14.0
xlrd==2.0.1
XlsxWriter==3.1.2
zipp==3.9.0
